requirements.txt
asgiref==3.3.1
Django==3.1.5
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
Markdown==3.3.3
PyJWT==2.0.1
python-memcached==1.59
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
